chore: moved App.kt out of ebms folder, and into the root project folder for ease of discovery
