+++ ebms-send-in/src/main/kotlin/no/nav/emottak/frikort/FrikortClient.kt
@@ -1,8 +1,8 @@
 package no.nav.emottak.frikort
 
 import no.nav.emottak.cxf.ServiceBuilder
-import no.nav.emottak.ebms.log
 import no.nav.emottak.fellesformat.FellesFormatXmlMarshaller
+import no.nav.emottak.log
 import no.nav.emottak.utils.environment.getEnvVar
 import no.nav.emottak.utils.environment.getSecret
 import no.nav.tjeneste.ekstern.frikort.v1.FrikortV1Port
