+++ ebms-send-in/src/test/kotlin/no/nav/emottak/PayloadIntegrationTestFelles.kt
@@ -8,7 +8,6 @@ import io.micrometer.prometheus.PrometheusConfig
 import io.micrometer.prometheus.PrometheusMeterRegistry
 import no.nav.emottak.auth.AZURE_AD_AUTH
 import no.nav.emottak.auth.AuthConfig
-import no.nav.emottak.ebms.ebmsSendInModule
 import no.nav.emottak.util.EventRegistrationServiceFake
 import no.nav.security.mock.oauth2.MockOAuth2Server
 import okhttp3.mockwebserver.MockResponse
